Split xilem lib.rs into multiple files
This commit only moves code around and changes some imports.
There are no changes affecting logic.
